Improve collaborative process initialization.
When a user is working on a model in collaborative editing mode and has
made changes that he hasn't saved yet (cache data), then, when another
user comes to work on the same file, by this commit, the content of the
new user editor will be synchronized with the content of the first one.
In other words, in the second editor, the user will have the content of
the first editor, not the original file content.